Added white wood background, css navigation bar
Boostrap folder for the background image template, and the beginning of the identify.php file. This file has an error when it comes the SQL statement to grab a column of the database as an array, as well as the return value of the compare face compare method. The identify.php file needs to be debugged.